# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pesisir Project",
        "description": "API Documentation for Pesisir",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user and get API token",
                "description": "Authenticate a user with email and password to receive an access token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@pesisir.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "man3sleman"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|XyZAbcDefGhIjKlMnOpQrStUvWxYz1234567890"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin User"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@pesisir.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Invalidate the current access token",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/complaints": {
            "get": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Get list of complaints",
                "operationId": "0888e2c7fc7dfed9d2cd561011ade1dd",
                "responses": {
                    "200": {
                        "description": "List of complaints with location"
                    }
                }
            },
            "post": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Create a new complaint",
                "description": "Store a newly created complaint with location and image",
                "operationId": "552e12061bd9cb4b042f2ed9c1735c33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "address",
                                    "phone",
                                    "desc",
                                    "date",
                                    "image",
                                    "loc_name",
                                    "loc_address",
                                    "lat",
                                    "long"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Sudirman No. 123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Illegal trash dumping"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-23"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "loc_name": {
                                        "type": "string",
                                        "example": "Kelurahan A"
                                    },
                                    "loc_address": {
                                        "type": "string",
                                        "example": "Jl. Kenanga, RT 01 RW 02"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -7.8014
                                    },
                                    "long": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 110.3647
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Complaint created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/complaints/{id}": {
            "get": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Get a single complaint",
                "operationId": "d63bd2b4f1f4514c3391875c72f8810a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Complaint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Complaint details"
                    },
                    "404": {
                        "description": "Complaint not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Complaints"
                ],
                "summary": "Delete a complaint",
                "operationId": "b4654fb38217112b6c599a681917b596",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Complaint ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Complaint deleted"
                    },
                    "404": {
                        "description": "Complaint not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}